fix "There should be no puts/print and gets.chomp outside main.rb file."
