+++ lib/player.rb
@@ -1,0 +1,8 @@
+class Player
+  attr_reader :name, :marker
+
+  def initialize(name, marker)
+    @name = name
+    @marker = marker
+  end
+end
